feat(gmail): NormalisePipeline → NormaliseWalk on fixed input
R5 decompose, commit 2/3. Replace NormalisePipeline (4-scenario picker
+ stepper, two verbs in the controls slot) with NormaliseWalk: one
verb (step), fixed input J.Ohn.Doe+promo@gmail.com, four steps
(typed → lowercase → strip-dots → drop-tag). Workspace exception lives
in the prose <Aside>, not as a 4th chip — voice-profile §6 (callouts
own "the one rule to remember") and interactive-components.md §7
(don't ship a widget when the claim can be taught in prose).

Updates §3 prose and the <Aside> to match the new control surface
(remove the "Switch between the four inputs" / "fourth scenario"
references that no longer have a referent).
