+++ app/posts/how-gmail-knows-your-email-is-taken/widgets/index.ts
@@ -1,5 +1,5 @@
 export { TypingPause } from "./TypingPause";
-export { NormalisePipeline } from "./NormalisePipeline";
+export { NormaliseWalk } from "./NormaliseWalk";
 export { CacheWalk } from "./CacheWalk";
 export { BloomProbe } from "./BloomProbe";
 export { SignupRace } from "./SignupRace";
